// Jenta/types.ts
export type AspectRatio = '1:1' | '16:9' | '9:16' | '4:3' | '3:4';

export interface CustomizationOptions {
  lightingStyle: LightingStyle;
  cameraPerspective: CameraPerspective;
}

export interface ImageFile {
  base64: string;
  mimeType: string;
  name: string;
}

export interface HistoryItem {
  image: ImageFile;
  prompt: string;
}

export interface ProjectBase {
  id: string;
  name: string;
}
// ...
export interface StoryboardScene {
    id: string;
    sequence: number;
    description: string;
    cameraAngle: string;
    visualPrompt: string;
    image: ImageFile | null;
    isLoading: boolean;
    error: string | null;
}

export interface StoryboardStudioState {
    subjectImages: ImageFile[];
    customInstructions: string;
    aspectRatio: AspectRatio;
    scenes: StoryboardScene[];
    gridImage: ImageFile | null;
    isGeneratingPlan: boolean;
    isGeneratingGrid: boolean;
    isUploading: boolean;
    error: string | null;
}

export interface StoryboardStudioProject extends ProjectBase, StoryboardStudioState {}